requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyright==1.1.404
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rich==14.0.0
rich-toolkit==0.14.7
ruff==0.12.11
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
